+++ backend/contracts/YcRegistry.sol
@@ -29,10 +29,16 @@ contract YcRegistry is AccessControl {
     event StrategyAdded(IYcStrategy strategy);
     event AccountCreated(address indexed owner, IYcStrategy strategy, uint usdcAmount, uint ethAmount);
     event StrategyNetAPYsUpdated(address bot, IYcStrategy strategy);
+    event EthFixedReallocationFeeSet(uint128 oldEthFixedReallocationFee, uint128 newEthFixedReallocationFee);
+    event UsdcYieldFeeRateSet(uint16 oldUsdcYieldFeeRate, uint16 newUsdcYieldFeeRate);
+    event YctMinted(address indexed owner, uint amount);
+    event StrategyVaultAdded(IYcStrategy strategy, IVaultV2 vault);
 
     //----- STATE VARIABLES -----//
 
     bytes32 public constant BOT_ROLE = keccak256("BOT_ROLE");
+
+    bytes32 public constant ACCOUNT_ROLE = keccak256("ACCOUNT_ROLE");
 
     IYcFactory public factory;
 
@@ -47,14 +53,21 @@ contract YcRegistry is AccessControl {
 
     mapping(IYcStrategy => bool) public allowedStrategies;
 
+    // Packing : 128 + 16 = 144
+    uint128 public ethFixedReallocationFee;
+    /// @dev Scaled by 10^2 (i.e., 2 decimals).
+    uint16 public usdcYieldFeeRate;
+
     //----- FUNCTIONS -----//
     
-    constructor(ERC20 _usdc) {
+    constructor(ERC20 _usdc, uint128 _ethFixedReallocationFee, uint16 _usdcYieldFeeRate) {
         _grantRole(DEFAULT_ADMIN_ROLE, msg.sender);
 
         usdc = _usdc;
         factory = new YcFactory();
         yct = new YcToken();
+        ethFixedReallocationFee = _ethFixedReallocationFee;
+        usdcYieldFeeRate = _usdcYieldFeeRate;
     }
 
     /// @notice Adds a strategy to the registry.
@@ -73,12 +86,13 @@ contract YcRegistry is AccessControl {
     /// Transfers ETH to the account.
     /// Attempts to Transfer USDC to the highest performing yield vault according to the selected strategy.
     /// Mints 1 YCT and transfers it to the account.
-    function createAccount(IYcStrategy _strategy, uint _amount) payable external returns (IYcAccount) {
+    function createAccount(IYcStrategy _strategy, uint _amount, uint32 _noReallocationPeriod) payable external returns (IYcAccount) {
         require(allowedStrategies[_strategy], NotAllowedStrategy());
         require(address(accounts[msg.sender]) == address(0), AccountAlreadyExists());
         
-        IYcAccount account = factory.createAccount(usdc, _strategy, msg.sender);
+        IYcAccount account = factory.createAccount(usdc, _strategy, msg.sender, _noReallocationPeriod);
         accounts[msg.sender] = account;
+        _grantRole(ACCOUNT_ROLE, address(account));
 
         if (msg.value > 0) {
             (bool success, ) = address(account).call{value: msg.value}("");
@@ -92,6 +106,7 @@ contract YcRegistry is AccessControl {
 
         yct.mint(address(account), 10**18);
 
+        emit YctMinted(msg.sender, 10**18);
         emit AccountCreated(msg.sender, _strategy, _amount, msg.value);
 
         return account;
@@ -109,12 +124,42 @@ contract YcRegistry is AccessControl {
         return _revokeRole(BOT_ROLE, _account);
     }
     
+    /// @notice Adds a vault to a strategy.
+    /// This function can only be called by admin.
+    function addStrategyVault(IYcStrategy _strategy, IVaultV2 _vault) external onlyRole(DEFAULT_ADMIN_ROLE) {
+        _strategy.addVault(_vault);
+
+        emit StrategyVaultAdded(_strategy, _vault);
+    }
+
     /// @notice Update the net APY of the vaults for a specified strategy.
-    /// This function can only be called by tan authorized bot.
-    function updateStrategyVaultsNetAPY(IYcStrategy _strategy, IVaultV2[] memory _vaults, uint[] memory _vaultsNetApy) external onlyRole(BOT_ROLE) {
+    /// This function can only be called by an authorized bot.
+    function updateStrategyVaultsNetAPY(IYcStrategy _strategy, IVaultV2[] memory _vaults, uint32[] memory _vaultsNetApy) external onlyRole(BOT_ROLE) {
         _strategy.updateVaultsNetAPY(_vaults, _vaultsNetApy);
 
         emit StrategyNetAPYsUpdated(msg.sender, _strategy);
+    }
+
+    /// @notice Set the ETH fixed reallocation fee used when a bot reallocate USDC from an account.
+    function setEthFixedReallocationFee(uint128 _ethFixedReallocationFee) external onlyRole(DEFAULT_ADMIN_ROLE) {
+        (ethFixedReallocationFee, _ethFixedReallocationFee) = (_ethFixedReallocationFee, ethFixedReallocationFee);
+
+        emit EthFixedReallocationFeeSet(_ethFixedReallocationFee, ethFixedReallocationFee);
+    }
+
+    /// @notice Set the USDC Yield Fee rate used when a bot reallocate USDC from an account.
+    function setUsdcYieldFeeRate(uint16 _usdcYieldFeeRate) external onlyRole(DEFAULT_ADMIN_ROLE) {
+        (usdcYieldFeeRate, _usdcYieldFeeRate) = (_usdcYieldFeeRate, usdcYieldFeeRate);
+
+        emit UsdcYieldFeeRateSet(_usdcYieldFeeRate, usdcYieldFeeRate);
+    }
+
+    /// @notice Mints 1 YCT and transfers it to the account.
+    /// This function can only be called by an account.
+    function mintYct() external onlyRole(ACCOUNT_ROLE) {
+        yct.mint(address(msg.sender), 10**18);
+
+        emit YctMinted(msg.sender, 10**18);
     }
 
     receive() payable external {