+++ frontend/src/hooks/account/useGetAccount.ts
@@ -1,11 +1,21 @@
 import { contractABI as accountContractABI } from "@/constants/contracts/account";
 import { contractABI as ivaultv2ContractABI } from "@/constants/contracts/ivaultv2";
+import { contractAddress as registryContractAddress, contractABI as registryContractABI } from "@/constants/contracts/registry";
 import { contractAddress as yctContractAddress, contractABI as yctContractABI } from "@/constants/contracts/yct";
 import { contractAddress as usdcContractAddress, contractABI as usdcContractABI } from "@/constants/contracts/usdc";
 import { useBalance, useReadContract } from "wagmi";
 import { AccountType } from "@/types/AccountType";
 
 export const useGetAccount = (contractAddress: any) => {
+    const {data: usdcYieldFeeRate, isLoading: usdcYieldFeeRateIsLoading, isSuccess: usdcYieldFeeRateIsSuccess} = useReadContract({
+        address: registryContractAddress,
+        abi: registryContractABI,
+        functionName: "usdcYieldFeeRate",
+        query: {
+            enabled: !!contractAddress
+        }
+    });
+    
     const {data: currentVaultAddress, isLoading: currentVaultAddressIsLoading, isSuccess: currentVaultAddressIsSuccess} = useReadContract({
         address: contractAddress,
         abi: accountContractABI,
@@ -33,7 +43,7 @@ export const useGetAccount = (contractAddress: any) => {
         }
     });
 
-    const {data: shares, isLoading: sharesIsLoading, isSuccess: sharesIsSuccess} = useReadContract({
+    const {data: shares, isLoading: sharesIsLoading, isSuccess: sharesIsSuccess, error} = useReadContract({
         address: currentVaultAddress as `0x${string}`,
         abi: ivaultv2ContractABI,
         functionName: "balanceOf",
@@ -49,7 +59,7 @@ export const useGetAccount = (contractAddress: any) => {
         functionName: "previewRedeem",
         args: [shares],
         query: {
-            enabled: !!currentVaultAddress && !!shares
+            enabled: sharesIsSuccess
         }
     });
 
@@ -58,6 +68,24 @@ export const useGetAccount = (contractAddress: any) => {
         abi: usdcContractABI,
         functionName: "balanceOf",
         args: [contractAddress],
+        query: {
+            enabled: !!contractAddress
+        }
+    });
+
+    const {data: capital, isLoading: capitalIsLoading, isSuccess: capitalIsSuccess} = useReadContract({
+        address: contractAddress,
+        abi: accountContractABI,
+        functionName: "capital",
+        query: {
+            enabled: !!contractAddress
+        }
+    });
+
+    const {data: depositAmount, isLoading: depositAmountIsLoading, isSuccess: depositAmountIsSuccess} = useReadContract({
+        address: contractAddress,
+        abi: accountContractABI,
+        functionName: "depositAmount",
         query: {
             enabled: !!contractAddress
         }
@@ -80,11 +108,14 @@ export const useGetAccount = (contractAddress: any) => {
         }
     });
 
-    const totalUsdc = (usdcBalance ? Number(usdcBalance) : 0) + (assets ? Number(assets) : 0);
-
+    const capitalUsdc = (usdcBalance ? Number(usdcBalance) : 0) + (capital ? Number(capital) : 0);
+    const usdcFee = ((assets ? Number(assets) : 0) - (depositAmount ? Number(depositAmount) : 0)) * (usdcYieldFeeRate ? Number(usdcYieldFeeRate) : 0) / 10**5;
+    const earnedUsdc = (assets ? Number(assets) : 0) - (capital ? Number(capital) : 0) - usdcFee;
+    
     let account: AccountType = {
         address: contractAddress,
-        usdc: totalUsdc,
+        capitalUsdc: capitalUsdc,
+        earnedUsdc: earnedUsdc,
         yct: yctBalance ? Number(yctBalance) : 0,
         eth: eth?.value ? Number(eth?.value) : 0,
         noReallocationPeriod: noReallocationPeriod ? Number(noReallocationPeriod) : 0,
@@ -95,8 +126,10 @@ export const useGetAccount = (contractAddress: any) => {
     return {
         data: account,
         isLoading: currentVaultAddressIsLoading || strategyAddressIsLoading || sharesIsLoading || assetsIsLoading || usdcBalanceIsLoading 
-            || yctBalanceIsLoading || ethIsLoading || noReallocationPeriodIsLoading,
+            || yctBalanceIsLoading || ethIsLoading || noReallocationPeriodIsLoading || capitalIsLoading || depositAmountIsLoading
+            || usdcYieldFeeRateIsLoading,
         isSuccess: currentVaultAddressIsSuccess && strategyAddressIsSuccess && sharesIsSuccess && assetsIsSuccess && usdcBalanceIsSuccess
-            && yctBalanceIsSuccess && ethIsSuccess && noReallocationPeriodIsSuccess
+            && yctBalanceIsSuccess && ethIsSuccess && noReallocationPeriodIsSuccess && capitalIsSuccess && depositAmountIsSuccess
+            && usdcYieldFeeRateIsSuccess
     };
 };