+++ frontend/src/types/EventType.ts
@@ -2,5 +2,6 @@ export interface EventType {
     eventName: string,
     transactionHash: string,
     blockNumber: bigint,
+    order: number,
     args: {}
 }