+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [User] View history
 - [Authorized bot] Update yield data
 - [Bot] Check if an account need reallocation
+- [Bot] Reallocate tokens of an account
 - [Admin] Add an authorized bot (backend only)
 - [Admin] Remove an authorized bot (backend only)
 - [Admin] Add a vault in a strategy (backend only)